category types

+++ backend/src/Controller/Finance/Category/CategoryController.php
@@ -6,6 +6,7 @@ namespace App\Controller\Finance\Category;
 
 use App\Controller\AbstractApiController;
 use App\Entity\Category;
+use App\Enum\CategoryType;
 use App\Repository\CategoryRepository;
 use App\Request\Category\CreateCategoryRequest;
 use App\Request\Category\UpdateCategoryRequest;
@@ -23,9 +24,21 @@ class CategoryController extends AbstractApiController
     }
 
     #[Route('', name: 'list', methods: Request::METHOD_GET)]
-    public function list(): JsonResponse
+    public function list(Request $request): JsonResponse
     {
-        return $this->respond($this->categoryRepository->findBy([], ['name' => 'ASC']));
+        $criteria = [];
+
+        $userCategoriesOnly = (bool) $request->query->get('userCategoriesOnly', '1');
+        if ($userCategoriesOnly) {
+            $criteria = ['type' =>
+                [
+                    CategoryType::USER,
+                    CategoryType::UNCATEGORIZED,
+                ]
+            ];
+        }
+
+        return $this->respond($this->categoryRepository->findBy($criteria, ['name' => 'ASC']));
     }
 
     #[Route('/{category}', name: 'get', methods: Request::METHOD_GET)]
@@ -40,6 +53,7 @@ class CategoryController extends AbstractApiController
         $category = (new Category())
             ->setName($request->getName())
             ->setColor($request->getColor())
+            ->setType(CategoryType::USER)
         ;
 
         $this->categoryRepository->save($category);
@@ -63,6 +77,10 @@ class CategoryController extends AbstractApiController
     #[Route('/{category}', name: 'remove', methods: Request::METHOD_DELETE)]
     public function remove(Category $category): JsonResponse
     {
+        if (!$category->isDefinedByUser()) {
+            throw $this->createAccessDeniedException();
+        }
+
         $this->categoryRepository->remove($category);
 
         return $this->respond(new EmptyResponse());
